check authorization for inviting/deleting member

+++ server/controllers/fridges_users.js
@@ -43,15 +43,24 @@ const addUserToFridgeByEmail = async (req, res) => {
     try {
         const { email } = req.body
         const fridge_id = parseInt(req.params.fridge_id)
+
+        const manager = await pool.query('SELECT user_id FROM fridges WHERE id = $1', [fridge_id])
+        if ((!req.user) || manager.rows[0].user_id !== req.user.id) {
+            res.status(409).json( { error: `User ${req.user.id} is not the manager of fridge ${fridge_id}` } )
+            return
+        }
+
         const user = await pool.query('SELECT * FROM users WHERE email = $1', [email])
         if(user.rows.length === 0) {
             res.status(409).json( { error: `User ${email} does not exist` } )
-        } else {
-            const user_id = user.rows[0].id
-            const results = await pool.query('INSERT INTO fridges_users (fridge_id, user_id) VALUES ($1, $2) RETURNING *', [fridge_id, user_id])
-            res.status(201).json(results.rows[0])
-            console.log(`User ${email} added to fridge ${fridge_id}`)
+            return
         }
+
+        const user_id = user.rows[0].id
+        const results = await pool.query('INSERT INTO fridges_users (fridge_id, user_id) VALUES ($1, $2) RETURNING *', [fridge_id, user_id])
+        res.status(201).json(results.rows[0])
+        console.log(`User ${email} added to fridge ${fridge_id}`)
+
     } catch (error) {
         res.status(409).json( { error: error.message } )
     }
@@ -61,6 +70,11 @@ const deleteUserFromFridge = async (req, res) => {
     try {
         const { user_id } = req.body
         const fridge_id = parseInt(req.params.id)
+        const manager = await pool.query('SELECT user_id FROM fridges WHERE id = $1', [fridge_id])
+        if ((!req.user) || manager.rows[0].user_id !== req.user.id) {
+            res.status(409).json( { error: `User ${req.user.id} is not the manager of fridge ${fridge_id}` } )
+            return
+        }
         const results = await pool.query('DELETE FROM fridges_users WHERE fridge_id = $1 AND user_id = $2', [fridge_id, user_id])
         res.status(200).json(results.rows)
     } catch (error) {
